fix argument size

+++ libfreeipmi/api/ipmi-oem-intel-node-manager-cmds-api.c
@@ -313,9 +313,9 @@ ipmi_cmd_oem_intel_node_manager_set_node_manager_alert_thresholds (ipmi_ctx_t ct
                                                                    uint8_t target_lun,
                                                                    uint8_t domain_id,
                                                                    uint8_t policy_id,
-                                                                   uint8_t *alert_threshold1,
-                                                                   uint8_t *alert_threshold2,
-                                                                   uint8_t *alert_threshold3,
+                                                                   uint16_t *alert_threshold1,
+                                                                   uint16_t *alert_threshold2,
+                                                                   uint16_t *alert_threshold3,
                                                                    fiid_obj_t obj_cmd_rs)
 {
   fiid_obj_t obj_cmd_rq = NULL;
