+++ libfreeipmi/include/freeipmi/api/ipmi-oem-intel-node-manager-cmds-api.h
@@ -90,9 +90,9 @@ int ipmi_cmd_oem_intel_node_manager_set_node_manager_alert_thresholds (ipmi_ctx_
                                                                        uint8_t target_lun,
                                                                        uint8_t domain_id,
                                                                        uint8_t policy_id,
-                                                                       uint8_t *alert_threshold1,
-                                                                       uint8_t *alert_threshold2,
-                                                                       uint8_t *alert_threshold3,
+                                                                       uint16_t *alert_threshold1,
+                                                                       uint16_t *alert_threshold2,
+                                                                       uint16_t *alert_threshold3,
                                                                        fiid_obj_t obj_cmd_rs);
 
 int ipmi_cmd_oem_intel_node_manager_get_node_manager_alert_thresholds (ipmi_ctx_t ctx,
